pyproject.toml
[tool.poetry]
name = "sync_service"
version = "0.1.0"
description = "File synchronization service"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
watchdog = "^2.1.6"
PyYAML = "^6.0"
tqdm = "^4.64.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
black = "^22.3.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dsync = "sync_service.cli.main:main" 